IP Masking: Enable for cswiki beta
Why:

The Growth team's features are fully activated
only on a couple of beta wikis (such as cswiki beta),
which makes them harder to test on dewiki.

Enabling IP Masking on an additional wiki is easier
than ensuring Growth features work well on dewiki beta,
especially since production dewiki doesn't have all
the features. Enabling IP Masking on cswiki beta
will allow Growth to prepare for IP Masking better.

In addition of being helpful to the Growth team,
having a second wiki with IP Masking enabled
will also make it possible to test IP Masking
cross-wiki.

Since prior to this patch, IP Masking was only
active on dewiki beta, this patch also introduces
several changes that are needed because of that
(most notably, it changes the serial provider
from local to centralauth).

While at it, the patch also ensures temporary accounts
behave as temporary accounts even on (beta) wikis
where IP Masking is not enabled.

For easier control (and because IP Masking is now enabled
on more than a single wiki), enrollment to IP Masking is
moved to a custom IS-labs.php variable, called
`wmgEnableIPMasking`.

What:
* Set CentralAuth as the serial provider (requires
  `global_user_autocreate_serial` to exist in
  the `centralauth` database).
* Set scramble as the serial mapping
* Add `wmgEnableIPMasking` to IS-labs, which
  controls whether IP Masking is enabled.
* Enable IP masking on cswiki beta

Notes:
Because this patch changes the serial provider, a SQL
query needs to be executed shortly before deploying
the patch:

```
  INSERT INTO centralauth.global_user_autocreate_serial (uas_shard, uas_value)
  WITH RECURSIVE _range(i) AS (SELECT 0 AS i UNION ALL SELECT i+1 FROM _range WHERE i < 7) -- for i in 0..7
  SELECT i, (SELECT uas_value FROM dewiki.user_autocreate_serial) + 100 + i FROM _range;
```

Bug: T342034
Change-Id: I76036e4e0b5877fde59e1f18f14e4c5fc989e048