Profiler: Remove "toobig" filter from Arc Lamp ingestion
As part of developing Excimer UI (with Speedscope), Tim updated
php-excimer to make truncated traces easier to detect and more useful
to keep.

With php-excimer 1.1.1 (deployed last month), Excimer inserts
"excimer_truncated" as the top frame if it was truncated due to
being too deep.

Together with the logic we already have to ensure the first frame
is set to the PHP entry point, this should avoid most ways that
flame graphs can become noisy or confusing.

This is motivated by simplifying Profiler.php so as to be closer
in sync with the recommended way we promote ArcLamp.php at
https://gerrit.wikimedia.org/g/performance/arc-lamp. This stood
out as something to either add there, or remove here.

Bug: T337873
Change-Id: Idada3761342c17a4c4c14e1584d084ad7d242790