api: Fix broken /api/index.html rendering
The references to the Swagger stylesheet no longer existed, as both
the stylesheet location, and the accompanying markup format, have
changed a lot over the years. To the point that the file is now a
404 Not Found, and its replacement would not actually make the page
much better.

https://phabricator.wikimedia.org/F37150701

Convert to semantic HTML markup, and inline two or three basic
style rules, akin to our error pages.

Bug: T113114
Change-Id: If3db955f9adadf0f621433423eef929c34f72145