ext-ORES: force cswiki to use the ORES settings/backend
The cswiki community informed us that a big number of
edits in RecentChances are marked as "very likely bad faith",
when in reality their score says the opposite. It seems
the same issue happened with fiwiki, probably a corner
case in the way the extension evaluates model thresholds
for certain wikis.

Revert to the original settings while the ML team investigates
what is the issue.

Bug: T343308
Change-Id: I27d170fb4a7e40948eb416ee31eb8aae474494b5