ores-extension: fix thresholds
Some thresholds need to reversed (substracted from 1). This happens for thresholds in the false class of each model.
The evidence for this lies withing the code of the extension.
https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/extensions/ORES/+/refs/heads/master/includes/ThresholdParser.php#97
When the extension parses thresholds responses as returned from ORES it has the following check:
elseif ( $bound === 'max' && isset( $statsData['false'][$stat]['threshold'] ) ) {
			$threshold = $statsData['false'][$stat]['threshold'];
			$threshold = 1 - $threshold;

Bug: T343308
Change-Id: Id6b7cb5452b5048fc0e3b14ed764a79802f9509c