Add new OAuth Rate Limiter tier for Wiki Education
The Wiki Education project needs a higher threshold for their
API Gateway's token rate limit to support a bulk-importer
application called "Impact Visualizer" that calls Lift Wing
(more info in the task).

Using "internal" seems to be counter-intuitive, create a new
ad-hoc tier for them is more clear (that we can also change
according to the needs of the use case).

Bug: T345394
Change-Id: I5749a19e7fbc450cb183f958c642311620b496e8