Add /.well-known/apple-developer-merchantid-domain-association
Why:
Per T346055, we need this file for FRUP integration for
donate.wikipedia.org.

donate.wikiPedia.org (stress mine using capital P) is an Apache
ServerAlias for donate.wikiMedia.org (stress again mine), which has a
docroot of docroot/wikimedia.org, which is a symlink to the
standard-docroot.

What:
Instead of putting this file under the standard-docroot, following the
same approach we have for thankyou.wikipedia.org and give donate it's
own docroot. We 'll need to also alter the apache configuration of
course

Bug: T346055
Change-Id: I4354f17f027a449f74c1aea7e3be1bf9074e2f0f