Allow FundraiseUp scripts in Donatewiki CSP
Fundraising is experimenting with a third-party fundraising platform
which needs to load scripts only on donatewiki (thus under donor
privacy policy, not reader privacy policy). You can load the form and
see all the CSP warnings by going to
https://donate.wikimedia.org/?withJS=MediaWiki:FundraiseUp.js&form-template=FundraiseUp

Per the comment on this variable I've tagged two members of the
security team on the associated Phab.

Bug: T345379
Change-Id: I5bc4ebfd0c3d8eb7c77f9fd0e469f139f9c88906