logging: Remove useMicrosecondTimestamps(false) calls
This was added in 2015 due to costly overhead on HHVM and PHP < 7.1
for computing microseconds on DateTime objects.

However, on PHP 7.1 and later this is computed by default by PHP,
which means all it is doing now is add lots of dynamic calls for
each new getLogger() object to also call useMicrosecondTimestamps(false).

Per <https://github.com/Seldaek/monolog/commit/4a43d9b17c>,
this setting is effectively ignored by Monolog as of v2.0. The hack that
we put in place upstream in 2015 is never used on PHP 7.1 and later.
The only difference it makes still calling `format('Y-m-d\TH:i:s.uP')`
or `format('Y-m-d\TH:i:sP')`.

Bug: T116550
Change-Id: I68a87d20156c99f4d907400a5c3760fbeb5155c2