Stop writing to $wgCentralAuthCookieDomain in 'EnterMobileMode' hook
This may not even work, depending on when the 'EnterMobileMode'
handler is called and when CentralAuthSessionProvider reads and
caches $wgCentralAuthCookieDomain.

Keeping the value of $wgCentralAuthCookieDomain consistent with
the keys of $wgCentralAuthAutoLoginWikis allows us to simplify
some code in Ic1ac21c61e3d2cf87f1834a791a1c2c6e14d3cfa (and it's
already documented in CentralAuth extension.json that the keys
"should be set to the cookie domain name").

The 'WebResponseSetCookie' hook should suffice for T49647.

Change-Id: I47ebde47e758402b7bb9f7ab11241f0bd4eb5e36