ORES: Set default value of OresLiftWingAddHostHeader to true
When this is enabled the appropriate host header is added in the request
made to Lift Wing.

The reason of this change is that the config variable
OresLiftWingAddHostHeader has been set to false in the extension
as it is not required by instances that don't use the internal endpoint.

Explicitly setting it to true as required by the production instances
for all wikis as they use the internal Lift Wing endpoint.

Bug: T351703
Change-Id: I6669022d0f167b0da26510bbbe6e426da67fd49a