Remove partial migration of VisualEditorFeatureUse instrument
There is a partial migration of the VisualEditorFeatureUse instrument to
the MP client library. It uses dispatch(), which is now deprecated in
favour of submitInteraction(). We should remove the partial migration of
the instrument to minimise our own maintenance burden.

In this specific change we are removing the configuration for the stream
that was created for that migration.

Bug: T351337
Change-Id: I4cf07f1c8c1cc0efd3c3957343f231617a942011