cirrus: Enable event bus bridge on more wikis
This set of wikis has been chosen for the initial production
deployment of the new cirrus updater. The refresh events are
necessary for this updater to function.

When fully enabled this will be a mid-high volume event stream. Prior
to deployment we must verify that the topic has been configured
appropriately, tracked in T351503.

Bug: T352335
Change-Id: I934ff75ce817f1968be24cf8332bdc40ca2e5c3d