Add a testing stream for page-prediction-change events
The ML team is currently publishing events from Lift Wing to
Event Gate, including revscoring model and outlink-topic model.
However, when testing in staging, outlink-topic model uses the
prod stream.

This change adds a testing stream for prediction-change events.
The purpose is to be able to test the pipeline without interfering
with the production streams. This will allow us to test current
and future models.

Bug: T349919
Change-Id: I1349d617e1f1d2539eb111a8cf2213afe670af2a