Disable max width for index namespace
It's been requested that we disable fixed width in vector 2022 for the
index namespace.  Since index namespace numbers vary across wikis, we
handle the config change in CommonSettings as opposed to
InitialiseSettings, under where it was previously done with the page ns

Couldn't figure out how to test CommonSettings locally, so the current
test plan is to ensure this is working during the debug/staging step of
a config deploy.  Welcome any and all advice as always!

Bug: T352162
Change-Id: If739ca3d6398415a7d8c6247363884f0f4feeb69