Configure parser cache filters for parsoid-pcache
This sets up wgParserCacheFilterConfig to disable caching of parsoid
renderings for the file namespace on commons and entity namespaces on
Wikidata. We want to disable caching for these namespaces because they
are large and rarely viewed. They would flood the parser cache with
useless data.

This removes the wgTemporaryParsoidHandlerParserCacheWriteRatio which we
were keeping around in order to disable parsoid-pcache on Commons and
Wikidata.

NOTE: this should be tested manually on debug hosts by editing, purging
and viewing pages in the relevant namespaces. Viewing with useparsoid=1
should also be tested. We should also monitor the impact on parser cache write rates after deployment: https://grafana.wikimedia.org/goto/oKmSJAtIk?orgId=1

Bug: T346765
Bug: T355375
Change-Id: I9f8e18254fffd10ce7872f52c29b57dcc9bf59be