Profiler: Silence "RedisException: Connection timed out"
We already do most profiling async and catch all exceptions to prevent
impact on the web response. However, for ease of debugging in the
future, we do still report it as warning to Logstash.

Since timeouts are a known and acceptable thing to happen, silence
these to prevent inactionable noise in Logstash.

If timeouts were to become unacceptably common, we will find out via
the alert we have on the StatsD metric. We do not rely on MediaWiki
error logging to discover that, and its error logging provides no
details that we don't already know.

If an unknown/new issue were to happen, we do still log those, as its
details may be important to investigating such future issues.

Bug: T348756
Change-Id: Ib57c980034b3da1fc282386f610240cb2cd218ba