Enable $wgLocalHTTPProxy on group1 wikis
What:
Enable group1 wikis outside of mw-on-k8s to use the local envoy proxy
for mediawiki calls.

Why:
We want to switch mediawiki to call mw-api-int instead of the baremetal
appservers cluster once commons has been migrated to mw-on-k8s. To
support this, we need mediawiki to go through the envoy local proxy for
internal calls.

Bug: T298265
Change-Id: I8ebcf2c1b09cdb4229dd2e03de25e8e95024208d