Enable $wgLocalHTTPProxy on all wikis
What:
Use the local envoy proxy for mediawiki calls on all wikis. Use the
intermediate $wmgUseLocalHTTPProxy setting for easy rollback in case of
issue.

Why:
We want to switch mediawiki to call mw-api-int instead of the baremetal
appservers cluster once commons has been migrated to mw-on-k8s. To
support this, we need mediawiki to go through the envoy local proxy for
internal calls.

Bug: T298265
Change-Id: Ieb0a5586e080b42280e44d2559cb3ca143498302