cirrus: Transition remaining cloudelastic wikis to streaming updater
We have been running 10-25% of the cloudelastic write load exclusively
through the replacement streaming updater for a couple weeks now.
During that time we have validated the deployment and are ready for
the next stage.

This patch fully transitions the cloudelastic cluster off of the job
queue based updater. The dependent patch turns on writes from
the streaming updater.

Bug: T358518
Depends-On: I662174f783ad233a8c5b36066aa7b013e15fe4bc
Change-Id: Id039e422ba726bacfbb9c69f87cc73ab4055a297