Disable special pages on a per name basis
Follow up to fcd37c2.

Rather than disable night mode on all special pages, limit it to
pages. The list is generated from Special:SpecialPages for the
target projects.

This means that night mode is supported on Special:MobileOptions
(which is a requirement for night mode to work) and we are able
to enable the feature on special pages that support night mode.

Bug: T359183
Change-Id: Ib9e7788274fde341571a78b7b258d6a73fc7a981