tests: Fix PHP 8.2 warnings
> Creation of dynamic property InitialiseSettingsTest::$config is deprecated
> Creation of dynamic property NocFileServeTest::$nocConfDir is deprecated
> Creation of dynamic property NocHighlightTest::$nocConfDir is deprecated

> PHP Deprecated:
> str_replace(): Passing null to parameter #3 ($subject) of type array|string
> is deprecated in SiteConfiguration.php on line 282
>
>   at SiteConfiguration->siteFromDB($wiki = 'default')
>   at str_replace($search = '_', $replace = '-', $subject = NULL)

"default" is not a valid wiki, it is in the InitialiseSettings array
as the default for other wikis, it can't be usefully tested as a
wiki because it doesn't match any of the registered wgConf suffixes
and thus can't produce a site/lang pair. Rather than "fixing" the
source code with unwanted tolerance for this bug, fix the bug
at the source instead by skipping that key.

Change-Id: I533b5ca09d1c702413350062fe8abf08904c077c