wikitech: fix handling of Gerrit status code
The BlockIpComplete and UnblockUserComplete hooks would log a failure
whenever Gerrit returns a status code different than 204. That is the
code returned when an account is deleted, but other cases yields
different status codes:

409 if the user was already deactivated
201 when it is made active again
200 when it was already active

Handle the other status codes to prevent misleading logs.

Log the success, and log the account id in addition of the username.

Bug: T307558
Change-Id: Id080b1a4f1166680156947348fa2e5fb0c1743d6