Add wmgUseCommunityConfiguration
To make it possible to load CommunityConfiguration extension
in Wikimedia environment, declare the wmgUseCommunityConfiguration
variable.

This is marked as 'do not enable without Growth team OK', because
the extension is currently only meant to be enabled in Beta.

WMF Security completed an Application Security Review, labelling
the extension with a risk of Low, which means the risk is automatically
accepted according to the current WMF Risk Management Policy. Results
are available at T349568.

Bug: T357766
Change-Id: I969b8e4988878485cb654f7b5896850bdbd94b8e