Scope temp user reserved pattern to temp users
In 6c61321d0f840322f4d439a8a6912ab7240c1698 we set
AutoCreateTempUser['reservedPattern'] to prevent new accounts
being created starting with '~'. This is because new temp users
will start with '~'.

However, this prevents existing users with names starting with '~'
from autocreating accounts with their own name on other wikis.

Instead, prevent new accounts starting with '~2'. Note that this
is the actual pattern that will be recognized by MediaWiki as a
temporary user (meaning that we don't have to rename users
starting with '~').

Bug: T361021
Bug: T349506
Change-Id: Ie98c7d9fcbdc812b5d8b4abfba6cb38497513c09