Remove buffer on cirrussearch-request log channel
These events need to be produced to kafka in a timely manner, but
during maintenance scripts we see events produced long after the
timestamp they were generated at.  While we don't have strong proof,
suspect that those events are being buffered here in the logger and
flushed much later in the maintenance script run. In a typical
web request only a single cirrussearch-request log is ever generated,
and it's done during postsend. As such this should have no practical
effect on those requests.

Bug: T359580
Change-Id: I05633498762fd91052aadbae975288789a783b2b