logging: always register udp2log handlers
Since the initial change introducing Monolog (2680380cba) the udp2log
handler for a given level is only added as a handler when a channel
requires it.

With the intent to make `udp2log-error` the default fallback logger, I
need to ensure it exists which might not be the case when none of the
channel use `'udp2log' => 'error'`.  Since we already create and
register logstash handlers for all levels, do the same for udp2log.

That later let us add `udp2log-error` as a default handler without
having to create and register it.

Bug: T228838
Change-Id: I3b9ffcebc884b0c28c7e46b8f483b202b3cd4aaf