Set conflicting gadget settings for the Cite extension
The code responsible for Reference Previews is moving to the Cite
extension. There we also need to respect conflicting gadgets and
their names.

This patch makes sure that Cite/Reference Previews gets the gadget
name configuration set up while transitioning the settings. One of
the two will remain a copy of the other.

Bug: T362771
Change-Id: Ia028c41f8aaa1c522dfc7c372e1ce51e40933a5e