Disable ParserMigration on commonswiki
With Parsoid, structured data on files is not visible (but can still be
added, potentially promoting the addition of duplicate / redundant
statements); disable ParserMigration for now so that users don’t
experience this issue. We can hopefully revert this once T351113 has
figured out how to support multiple slots in Parsoid.

Bug: T364228
Change-Id: I7ca88556eac51ecbaf6cff2bd90db6bebd61beb4