Use encrypted Argon2 Hashes to store user passwords
We are currently using PBKDF2 for storing passwords.

Argon2i was designed to resist side-channel attacks and Argon2d to
resist GPU cracking attacks. Argon2id is a hybrid designed to resist
both type of attacks.

Argon2 should be prefered since it is harder to implement scalably in
hardware than SHA512 (and thus PBKDF2) is.

We also want to start encrypting the our password hashes given the risk
of a db leak (which we hope never happens). The Argon2id function
actually supports encrypting the password with a secret (see [1]), but
php's password_hash() does not support passing the secret. So we need to
use EncryptedPassword instead which is a bit clunky but not an issue.

This patch must be preceded by a patch that defines the secret in
PrivateSettings.php.

[1] - https://github.com/P-H-C/phc-winner-argon2#library

Bug: T150647
Bug: T216682
Change-Id: Ieba4321187ad2549b2cd083c5ff65b22e5210b67