cirrus: Shift 25% of public wikis writes in eqiad to replacement updater
This stops cirrus from performing writes to the elasticsearch clusters
in eqiad for a subset of public wikis. On initial deployment there may
be a few logged errors from jobs that were queued when the clusters
were writable, but run after this is deployed. As long as job queue lag
is minimal the logged errors from the jobqueue should be brief.

Bug: T363475
Change-Id: I9b2dcd67da09726650090a6aba41277e12e54a76
Related: Ib462406651690d0aa38a2d3fb6140f7ba12d93f2