Allow async (job queue based) chunked upload on all wikis
Previously this was only allowed on commons and testwikis. While
it is rare for large files to be uploaded on local wikis, there
are cases. For example, silent films that are public domain in
US but not their country of origin are allowed on english wikipedia
but not commons and are often quite long.

Having the some config on all wikis helps reduce complexity.

Bug: T364644
Change-Id: I394f60f297e4cff807284f70a6b96cf42ad98e93