Update tagline and wordmark of Persian Wikibooks
Apparently while introduction of Vector-2022 a former version of
Persian Wikibooks' logo was assumed as the latest version and used
to pick the wordmark instead of the 2013 version, this uses the
newer version's wordmark which was in use since 2013.

The new versions don't use SVG clip so as English versions unlike
the previous version of the files.

The SVG files are also uploaded in Wikimedia Commons by the designer
who has applied some path improvements and optimizations also:

* https://commons.wikimedia.org/wiki/File:Wikibooks-tagline-fa.svg
* https://commons.wikimedia.org/wiki/File:Wikibooks-wordmark-fa.svg

Bug: T365913
Change-Id: I25f857a6c0d2c418d7ac4c836d9fe18f4b76303d