Handle sso.wikimedia.org domain
Support the use of https://sso.wikipedia.org/<host>/...
login URLs which resolve to the current wiki, no matter
which wiki the user is on, but always see the same cookies.

This would be used to rely on a single central cookie during
login and signup, while still taking the configuration of the
user's wiki fully in account when rendering the login and
signup pages.

The immediate goal is to enable this setup on the beta
cluster; the decision whether to use it in production is
pending on further experimentation and discussion. But
much of the relevant code does not have a beta/production
split, so the patch adds support for both. In production,
the domain would not resolve; in beta, Idf70fc0afb46debc
adds support for it.
At this point no code is generating such URLs; that will
be done in CentralAuth in a follow-up (and enabled on
beta only, for now).

Details:
* Handle sso.wikimedia.org (or the beta equivalent) in
  multiversion so it resolves to the wiki indicated by
  the path prefix.
* Alter $wgServer, $wgCanonicalServer, $wgArticlePath,
  $wgScriptPath and $wgResourceBasePath to make URLs point
  to the SSO domain, not the normal wiki domain, and add
  the appropriate prefix. ($wgConf is unchanged so one can
  use WikiMap to get an URL that uses the normal wiki
  domain.)
  For $wgServer and $wgCanonicalServer (which are usually
  not used when generating URLs) it's not obvious whether
  this is a net positive.
* Alter $wgLoadScript so it uses the normal URL for the
  given wiki, to avoid splitting the ResourceLoader module
  cache; load.php ignores the session so there's no harm
  in this.
* Adjust session cookie settings so that the same cookie
  is used regardless of which wiki is requested via the
  SSO domain, and disable autologin while on that domain.
* Add sso.wikimedia.org to various domain allowlists.
* Set an URL variable, to be used by CentralAuth later.
* Show an error page if the SSO domain is used with a
  path specifying a non-SUL wiki.
* Make MWMultiVersion::getMediaWiki() testable, and add
  tests.

Security-wise, the following prevent the use of this domain
to access information from a private wiki:
* CommonSettings.php is hardcoded to exit when the SSO
  domain is used but the wiki is not using CentralAuth.
* Even if that were to be circumvented somehow, the private
  wiki would still use its usual authentication settings,
  so an unauthorized user would not be able to do anything
  they can't do on the private wiki's domain.
* CentralAuth hooks will disable any non-login-related
  functionality (such as normal pages, most special pages,
  most APIs). This will also prevent viewing normal content
  via the alternative URL and wasting cache space, causing
  SEO issues etc. This is done in I05be0088a262d.
See T367995 for further security considerations.

Soft-depends on Idf70fc0afb46debc32e72987da696abc23f1f65b.

Bug: T365162
Change-Id: I4a120d3024f655f8a1f020d980373ebd56b423bc
Depends-On: I05be0088a262d902d86c86300e4068ac943ad6a6