cirrus: Send weighted tags to known clusters
A recent commit, 29e5cf471d, stopped sending writes from cirrus
to the elasticsearch clusters in most circumstances. This worked
a little too well, and we need to bring back weighted_tags.

Private wikis are already excluded from this configuration, so
these are safe to send to cloudelastic as well. Related patch
is required for this to work and can be deployed in either order.

Related: I860b6ab054ff82c874cd1d0e6080377abf53d719
Change-Id: Ia29b4e3b02006691000ce69f8ddd0174d1c250c7