Set LinterParseOnDerivedDataUpdate to false
In preparation for If53841ee583ce240dd245d640b9ea9c97e1eaa55.

LinterParseOnDerivedDataUpdate should be false as long as we trigger
parsoid parses through the pregeneration mechanism in RESTbase. Once we
turn that off, LinterParseOnDerivedDataUpdate needs to be switched on to
ensure that lint data gets updated when pages are affected by template
changes.

Bug: T361013
Change-Id: Ib2c887380438fef5f5e5e997e11465254b54ad67