Growth: Use `growthexperiments` DB list for enabling GrowthExperiments
Prior to this patch, we enumerated the list of wikis in two places:

    (a) the `growthexperiments` dblist,
    (b) the `wmgUseGrowthExperiments` config variable

This was done for performance reasons, as loading a dblist used to have
performance implications (every new DB list used to meant having to load
a new file on every request). More recently, DB lists were reworked to make
them more performant. This was done by compiling the YAML definitions
to dblists-index.php file, which makes it possible to load all DBlists
for a wiki in constant time.

The Growth team is about to deploy CommunityConfiguration to all wikis
where GrowthExperiments is enabled. To make that possible, we would need
the data about the list of GE wikis in two other places.

Considering the performance implications are less important due to
the dblists-index refactor, this patch is adding the `growthexperiments`
list to MWMultiVersion and replacing the list of wikis in ext-GrowthExperiments
with the dblist. This patch also removes the `GrowthTest`, as it is no longer
important to assert (a) and (b) are in sync.

Bug: T364892
Change-Id: Ib95a14a5aa90c6b4d59465aa5fec91b2b57d99fa