Load EntitySchema on Test Wikidata clients
We need to load the extension, in client-only mode, on Wikibase client
wikis, to ensure that the data type is registered; otherwise, client
wikis may simply crash when trying to load any entity from the repo
where the data type has been used (even when loading unrelated
statements). This change does that for the Test Wikidata client wikis;
the real Wikidata client wikis will follow later (with T332157).

Also, use the 'wikidata' dblist when we want to change a setting both on
Wikidata and Test Wikidata (wikidata = wikidatawiki + testwikidatawiki).

Bug: T363153
Change-Id: Icf120c69f0da244160bb7df35c7f436c46d28342
Depends-On: I554aecb77c7c8401d28377d35789950b45082a24
Depends-On: I80ed2df297fd7161f0667e38effe0b51ad750987