wikidatawiki: Add namespace 640 (EntitySchema) to $wgContentNamespaces
The registration of the namespace was changed in EntitySchema change
Ia52f6c3d93 (commit 625905daa0), but the attempt to make it a content
namespace there fails. Work around it for now by making it a content
namespace in the config here, just like we already do for Property and
Lexeme (which really should also be declared content by Wikibase and
WikibaseLexeme automatically IMHO).

Bug: T368010
Change-Id: Icb2472b5d77b271b0ea5c25079105d0df24bf3dc