Add $wgMaxShellWallClockTime setting for shellbox
What:
Configure the child process timeout used in Shellbox.
(Shellbox is the only place that uses this value other than extensions
we don't use in production.)

Why:
PHP-FPM does not kill child processes when requests time out, so this
needs to be set to the same value as the timeout in the shellbox
deployment chart, otherwise we'll get orphaned long-running processes
that consume CPU but cannot return a result.

Bug: T356241
Change-Id: Ia79bb9af6f6378879db5280fe2c71b447c724856