Linter: trigger parsoid parses on template changes
This enables $wgLinterParseOnDerivedDataUpdate which triggers a parsoid
parse in order to update linter data whenever derived data updates are
happening. This is generally triggered by RefreshLinksJob.

The intent is to update linter data for pages affected by a template
change. So far, the necessary parsoid parse has been trigegred by a
request to the REST API priginating from RESTbase, following a no-cache
request from changeprop to RESTbase. This mechanism is about to be shut
off as a step in the sunsetting of RESTbase. To keep linter data up to
date, we need to enable this alternative mechanism.

Bug: T361013
Change-Id: I1118d89f6a3a3e44a9db0c9dab8740ce73755b24