Enable mul language code on Wikidata (limited mode)
On 29 July 2024, we want to roll out the `mul` language code to
Wikidata. This requires enabling it on all the client wikis as well
(otherwise they’ll probably start crashing when reading items with `mul`
labels), so for simplicity let’s just set the config to `true`
everywhere (it doesn’t hurt on wikis that don’t use Wikidata).

Note that wmgWikibaseTmpAlwaysShowMulLanguageCode remains unset on
Wikidata – the “full” release, when this is set to `true`, will only
come with T312176.

The Beta override is no longer needed and can be removed.

Bug: T330281
Change-Id: I7d91a0b0ee502e786946b13c7d925e3b7c1f6402