beta: Enable NetworkSession extension
Enables the NetworkSession extension and configures it with a
single user that is limited to read access (which any anonymous
user already has) and an auth token. Ip ranges are limited to
hosts in cloud not for security purposes, as it can only read
already public pages, but to verify operation of the extension.

Bug: T355267
Change-Id: I6cf3bad71744601db77507758cf879de2de47330