Move section mapping to separate file
This is a static array which never changes unless a new wiki is created
or in the past when a wiki got deleted (but we no longer delete wikis).

I would like to automate wiki creations further and automatically update
this mapping for the new wiki (which is usually going to go s5).

In order to make this more easy and robust, I would propose to move this
mapping to a separate file, which contains nothing else. First I had the
idea of just generating this on fly (Id029641034a4ec), but that is
probably not a good idea out of performance and stability reasons.

Doing the actual automation is going to happen in a follow-up patch.

I am also going to move the mapping of the centralauth db to s7 to
CommonsSettings, since it is a special thing, which should not be
hidden between the standard wiki mappings.

Change-Id: Ie326bab69fdf6c34e9a9f6f78d34036dbdc7c803