wikitech: de-wikitech mediawiki-config
Why:
We are migrating wikitech to mw-on-k8s, ie it
will become yet another wiki rather than the
snowflake it is now.

What:
Remove as many wikitech related configuration
overrides, one bit at a time.

This patch will be tested in debug hosts, but
will not be merged to production until we are
confident we have put all our ducks in a row.

Changes:
* don't load the ldap extention
* don't load developer block functionality
* use mcrouter instead of memcached-pecl
* use MainStash
* enable wikitech password reset (local db)
* use our generic jobqueue

Variables:
* wmgUseTimedMediaHandler
* wmgUseRestbaseVRS
* wmgUseParsoid
* wmgUseBounceHandler
* wgParserCacheType
* wgChronologyProtectorStash
* wmgUseMobileApp
* wgMFAutodetectMobileView
* wmgUseBounceHandler
* wgMessageCacheType
* wgSessionCacheType
* wgEchoSeenTimeCacheType
* wgCentralAuthSessionCacheType
* wgMWOAuthSessionCacheType
* wgMWOAuthNonceCacheType
* wgMainStash
* wgMicroStashType
* wgEnableEventBus
* wmgUseClusterJobqueue

Bug: T371537
Bug: T371592
Bug: T371374
Bug: T371359
Change-Id: I795d3bfb207de196bede700f397a2caee1af7a2b