Use encrypted PBKDF2 for wrapping B type passwords instead of Argon2
Wrapped passwords are represented as LayeredParameterizedPassword.

The core implementation of argon2 is not an instance of
ParameterizedPassword since it is already storing it's parameters (i.e.
the salt) in the hash itself.

But since it not an instance of ParameterizedPassword, when using it
to wrap password hashes, it no longer respects the user specific salt
while trying to validate the password, since validation happens by
cloning the password object an regenerating the hash and it doesn't
know the user specific salt when doing that.

So the password is never validated as correct and users are locked out
of their account.

So let us use PBKDF2 instead, since that is implemented as an instance
of ParameterizedPassword and we can easily use it.

Trying to write an ParameterizedPassword Argon2 implementation doesn't
really seem to be worth the effort for those old accounts since PBKDF2
is also quite secure.

Bug: T112359
Change-Id: I11a882ee6522c346dd38d1831971c62639cd4fae