Remove the "powered by mediawiki" override
This was needed during the transition period where mw's footer icon
would have looked weird but now that the new footer is deployed
everywhere, it's fine to remove the override.

Also remove the old png files, we kept them because of client side
caching or cdn caches but now enough time passed that we don't need them
anymore, may them rest in peace.

Change-Id: I90f3476e08771d1902f2a7864ebfa827be73981e