Enable dark mode for Creator: namespace in Commons
Wikimedia Commons actually doesn't have Portal namespace
but uses '100' for Creator: namespace so somehow an
surprising thing happened there that for two very
similar namespaces, Creator: (100) and Institution: (106)
we have a whole different behavior, e.g.:
* https://commons.wikimedia.org/wiki/Institution:Metropolitan_Museum_of_Art
* https://commons.wikimedia.org/wiki/Creator:Alan_Wilson

Which this removes the difference and makes the two to look
the same. Commons didn't have LiquidThreads either so no
namespace needs to be in its exceptions.

Change-Id: Ib89b07c65642327a59b49a661164090a51872322