Do not log failed autocreations on closed wikis as diagnostic errors
When a user which is not a steward goes to a closed wiki and an
automatic account creation happens, it is prevented since we do not want
to flood closed wikis with useless accounts.

This behaviour is an expected condition and not conflicting with a
decision made somewhere else.

Therefore, this is not an error and thus should be rather logged with
the level 'info'.

Bug: T373650
Change-Id: I0ff4d9cfbc074e791050e0cf382acfa4cb5169f6