Replace confusing uses of $wgDebugLogFile with $wmgExtraLogFile
With our configuration for $wgMWLoggerDefaultSpi and $wgLBFactoryConf,
setting $wgDebugLogFile to any value has no effect on MediaWiki.
(This is not documented anywhere, I had to check by reading the code.)
We are really just using it like a custom $wmg config setting, used
only by logging.php, but with a very confusing name.

Rename $wgDebugLogFile to $wmgExtraLogFile to make it more obvious it
doesn't affect MediaWiki.

Rename the 'wgDebugLogFile' Monolog handler to 'extraLogFile' as well.

Change-Id: Ieb8a1f20fc8a54f8349b107fe71d8d599d13f139