Remove labs settings for $wmgExtraLogFile that have no effect
The only code that reads $wmgExtraLogFile is in logging.php.
However, logging.php is loaded before CommonSettings-labs.php
(see CommonSettings.php, line 324 and 4412). So this did nothing.

That was probably for the best, the labs cluster might not be too
happy about extra couple of gigabytes of daily logs if this worked.

Remove the setting for $wmgDefaultMonologHandlers too. Because the
other config had no effect, the 'extraLogFile' handler was sending
all of the carefully formatted logs right to /dev/null (the default
value for $wmgExtraLogFile in production).

Change-Id: Id5eae91354ee37eaabe37ac1724785b23f798039