logging: Replace 'blackhole' handler with no handlers at all
Configuring no handlers should have the same effect as configuring
the 'blackhole' NullHandler, which does nothing.

By the way, NullHandler does not just do nothing, it also causes
other handlers for the same channel to do nothing. NoopHandler is
the one that just does nothing. We never combined 'blackhole' with
anything else, so this did not matter, but it seems scary to use it.
http://seldaek.github.io/monolog/doc/02-handlers-formatters-processors.html

Change-Id: I9bc5b5009b6586401f5875ca896cb8c651135b0b