cirrus: Introduce an expensive query pool counter
We have had a dedicated regex pool counter for some time, but
in a way that was simply our only very expensive full text query.
We are increasing the limits that we allow the deepcat query to
use which starts to push it into this expensive category of
queries as well.

Introduce a new ExpensiveFullText pool counter and give it the
same settings as the Regex variant. Related patch to cirrus
introduces the usage of the new name.

Bug: T369808
Related: I9f985ee5eca47a441aed885f9c0ec4ac45f59183
Change-Id: I39ead67b2e5b9f0bf53a196d6686a58661edf19d