Disable lead paragraph transform on Wikivoyages
On Wikivoyage's there are no infoboxes, so typically hoisting the lead
paragraph is applying in situations where an image begins with an image
rather than text. These are better addressed by editors switching the
order.

The lead paragraph doesn't apply in Parsoid which is enabled on all
wikis now so this makes the two experiences consistent

Change-Id: Iaf74e7f92dcc9b407a40714eb41cb37a989d3e25