Elevate pseudo-namespace MOS to a real namespace on most wikis which use it
**Since this change is potentially disruptive on enwiki (~2000 pages
on enwiki using the MOS pseudo-prefix) this patch omits enwiki from
the set of wikis touched.  Once we have validated this approach on the
smaller wikis, enwiki will be done in isolation as a followup
(Ia9dafebc9cfeb268db81fbe16165e16a9e97ee11).**

On 11 wikipedias (as/bn/dag/en/id/ja/ms/simple/sl/th/zh) the prefix
`MOS:` is used for shortcut links to [[Wikipedia:Manual Of Style]].
The MOS prefix conflicts with the language code for Mooré, so these
links will become interlanguage links when moswiki is added.  However,
namespace prefixes take precedence over interwiki prefixes, so creating
`MOS` as a "real" namespace will ensure that links on these wikis
won't break when `mos` is an interwiki/interlanguage link prefix.

A talk namespace needs to be setup to handle the existing pages
with titles of the form [[Talk:MOS:xxx]].

After deploying this configuration change, the maintenance script
needs to be run on the 11 affected wikipedias:

  php maintenance/run.php namespaceDupes.php --fix \
      --source-pseudo-namespace MOS --dest-namespace 126 \
      --move-talk

(Leave off the --fix option to do a dry run before actually writing
to the database.)

In order to move the [[Talk:MOS:xxx]] pages to [[MOS_talk:xxx]] you
also need to run:

  php maintenance/run.php cleanupTitles.php

(Add --dry-run to do a dry run before actually writing to the database.)

Bug: T363538
Change-Id: I91fa442aed595121dedfa7c07857301aff492996