Improve $wgFooterIcons override, simplify $wmgWikimediaIcon
`style="padding-left: 8px; padding-right: 8px;"` is not needed, that's
the default. `target="https://wikimediafoundation.org/"` is incorrect.
And the PNG assets defined in $wmgWikimediaIcon did not exist.
Apparently this was all a result of T256190.

Instead of writing out the HTML, which is prone to mistakes, define
the icon using the array syntax for $wgFooterIcons.

Change-Id: If0b8b1cca7362c25670ee7533668ec967f19bb0a